Added new tests for bad GET requests.  Fixed broken test.  Updated README file. Split out createLog/createReceipt into separate event-utils.js file.
